pyproject.toml
[project]
name = "andrew"
version = "0.0.1"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "mcp>=1.5.0",
    "openai>=1.68.2",
    "python-dotenv>=1.0.1",
]
